requirements.txt
discord.py==1.7.3
pydantic==1.8.2
python-dotenv
watchgod
slash-command